refactor: Main 컴포넌트 분리
