+++ src/App.js
@@ -1,12 +1,11 @@
 import "App.css";
 
-import CardList from "components/units/CardList";
-import Searchbar from "components/units/Searchbar";
-import Header from "components/units/Header";
-import Hero from "components/units/Hero";
-import Footer from "components/units/Footer";
 import { useEffect, useState } from "react";
 import { getUser } from "api/api";
+
+import Header from "components/units/Header";
+import Main from "components/units/Main";
+import Footer from "components/units/Footer";
 
 function App() {
   const [user, setUser] = useState();
@@ -22,17 +21,7 @@ function App() {
   return (
     <div>
       <Header user={user} />
-      <main>
-        <section className="hero">
-          <Hero />
-        </section>
-        <section>
-          <div className="contents">
-            <Searchbar />
-            <CardList />
-          </div>
-        </section>
-      </main>
+      <Main />
       <Footer />
     </div>
   );
