+++ src/components/units/CardList.js
@@ -1,23 +1,7 @@
-import { getFolder } from "api/api";
-import "components/units/CardList.css";
-import { useEffect, useState } from "react";
 import CardItem from "./CardItem";
+import "components/units/CardList.css";
 
-export default function CardList() {
-  const [links, setLinks] = useState();
-
-  useEffect(() => {
-    getFolderData();
-  }, []);
-
-  const getFolderData = async () => {
-    const result = await getFolder();
-    const { links } = result.folder;
-    setLinks(links);
-  };
-
-  console.log(links);
-
+export default function CardList({ links }) {
   return (
     <ul className="cards">
       {links?.map((link) => (
