+++ src/components/units/Main.css
@@ -1,0 +1,33 @@
+/* Hero */
+.hero {
+  text-align: center;
+  background-color: var(--color-primary-varient);
+}
+
+/* Contents */
+.contents {
+  padding: 2.5rem 2rem;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 2.5rem;
+  margin: auto;
+}
+
+/* Tablet */
+@media only screen and (min-width: 768px) and (max-width: 1199px) {
+  .contents {
+    padding: 2.5rem 2rem;
+  }
+}
+
+/* Mobile */
+@media only screen and (min-width: 375px) and (max-width: 767px) {
+  .contents {
+    width: 100%;
+    height: 100%;
+    padding: 1.25rem 2rem;
+    gap: 2rem;
+  }
+}
